--- src/pages/catalog/components/ScootersCatalogList/ScootersCatalogList.scss
@@ -1,11 +1,0 @@
-@import "../../../../variables";
-.scooters {
-  display: grid;
-  grid-column-gap: 30px;
-  grid-row-gap: 30px;
-  margin-bottom: 4rem;
-  @media screen and(min-width: $large-breakpoint) {
-    grid-template-columns: repeat(auto-fit, minmax(280px, 370px));
-    grid-template-rows: 480px;
-  }
-}
