+++ src/pages/scooters/CatalogPage.scss
