+++ src/pages/scooters/components/ScooterItem/ScooterItem.scss
@@ -1,0 +1,107 @@
+@import "../../../../variables";
+@import "../../../../mixins";
+.scooter-item {
+  display: grid;
+  grid-template-columns: 50% 50%;
+  &__image {
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: contain;
+    }
+  }
+  &__details {
+    padding: 2rem;
+  }
+  &__name {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+    margin-bottom: 1rem;
+    h1 {
+      font-size: $m-font-size-desktop;
+      font-weight: 600;
+    }
+    p {
+      font-size: $s-font-size-desktop;
+    }
+  }
+  &__colors {
+    display: flex;
+  }
+  &__color {
+    @include scooter-color;
+  }
+  &__description {
+    font-size: $xs-font-size-desktop;
+    letter-spacing: 2px;
+  }
+  &__scroll-to {
+    font-size: $xs-font-size-desktop;
+    display: block;
+    margin: 0.5rem 0;
+    text-decoration: underline;
+    background-color: transparent;
+    border: none;
+    font-weight: bold;
+    transition: color 0.15s linear;
+    &:hover {
+      color: $secondary-color;
+    }
+  }
+  .button {
+    margin-left: auto;
+  }
+  &__section-title {
+    width: 100%;
+    text-align: center;
+    grid-column: 1 / span 2;
+    padding: 2rem 0;
+    margin-top: 2rem;
+    background-color: rgba($secondary-color, 0.3);
+    h2 {
+      font-weight: 600;
+      font-size: $s-font-size-desktop;
+    }
+  }
+  &__specification-item {
+    padding: 2rem;
+    font-size: $xs-font-size-desktop;
+    font-weight: 600;
+    display: flex;
+    flex-direction: column;
+    &--title {
+      text-align: right;
+    }
+    &--value {
+      text-align: left;
+    }
+    span {
+      margin: 1rem 0;
+    }
+  }
+  &__shippings {
+    grid-column: 1 / span 2;
+    width: 80%;
+    margin: 2rem auto;
+    border: 1px solid black;
+    border-collapse: collapse;
+    th {
+      padding: 1rem 2rem;
+    }
+    td {
+      padding: 2rem 2rem;
+    }
+    th,
+    td {
+      text-align: center;
+      border: 1px solid black;
+      border-collapse: collapse;
+    }
+  }
+  #shipping,
+  #specification {
+    margin-top: -4rem;
+    padding-top: 4rem;
+  }
+}
