+++ src/pages/scooters/components/ScootersCatalogFilter/ScootersCatalogFilter.scss
