+++ src/pages/scooters/components/ScootersCatalogList/ScootersCatalogList.scss
@@ -1,0 +1,14 @@
+@import "../../../../variables";
+.scooters {
+  display: grid;
+  grid-column-gap: 30px;
+  grid-row-gap: 30px;
+  margin-bottom: 4rem;
+  @media screen and(min-width: $medium-breakpoint) {
+    grid-template-columns: 50% 50%;
+    grid-template-rows: 480px;
+  }
+  @media screen and(min-width: $large-breakpoint) {
+    grid-template-columns: 33% 33% 33%;
+  }
+}
