+++ src/pages/scooters/scooters.js
@@ -1,0 +1,435 @@
+import orange from "../../static/img/scooters/catalog/orange.png";
+import black from "../../static/img/scooters/catalog/black.png";
+import blueWhiteCustom from "../../static/img/scooters/catalog/blue-white-custom.png";
+import green2 from "../../static/img/scooters/catalog/green-2.png";
+import green from "../../static/img/scooters/catalog/green.png";
+import orange2 from "../../static/img/scooters/catalog/orange-2.png";
+import redWhite from "../../static/img/scooters/catalog/red-white.png";
+import redIvory125cc from "../../static/img/scooters/catalog/red-ivory-125cc.png";
+import RoyalAlloyGP200LCRed from "../../static/img/scooters/catalog/Royal-Alloy-GP-200-LC-Red.png";
+import UltraBlueIvory125cc from "../../static/img/scooters/catalog/Ultra-Blue-Ivory-125cc.png";
+import White2 from "../../static/img/scooters/catalog/White-2.png";
+import White from "../../static/img/scooters/catalog/White.png";
+
+export const DUMMY_SCOOTERS = [
+  {
+    name: "GT 50 AC",
+    price: "3455,99 zł",
+    image: orange,
+    colors: ["orange", "blue"],
+    rate: 24,
+    Manufacturer: "Forte",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "50cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 1,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "30 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "ARS 770S",
+    price: "5499 zł",
+    image: black,
+    colors: ["black", "blue"],
+    rate: 10,
+    Manufacturer: "Honda",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "80cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 2,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "35 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "HY 888Y",
+    price: "3799 zł",
+    image: blueWhiteCustom,
+    colors: ["blue", "white"],
+    rate: 16,
+    Manufacturer: "Moto Tech",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "100cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 3,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "40 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "TG 300S LC",
+    price: "3455,99 zł",
+    image: green2,
+    colors: ["green", "orange"],
+    rate: 12,
+    Manufacturer: "Mustang",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "125cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 4,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "45 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "TG 125S LC",
+    price: "9077 zł",
+    image: green,
+    colors: ["green", "blue"],
+    rate: 18,
+    Manufacturer: "Spark",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "150cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 5,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "50 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GP 125 LC",
+    price: "9899 zł",
+    image: orange2,
+    colors: ["orange", "blue"],
+    rate: 10,
+    Manufacturer: "Suzuki",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "50cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 6,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "33 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GP 200 LC",
+    price: "9998,99 zł",
+    image: redWhite,
+    colors: ["red", "white"],
+    rate: 11,
+    Manufacturer: "Viper",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "80cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 7,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "37 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "Custom Paint",
+    price: "11999 zł",
+    image: redIvory125cc,
+    colors: ["red", "blue"],
+    rate: 17,
+    Manufacturer: "Yiben",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "100cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 8,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "42 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GT 125i AC",
+    price: "3699 zł",
+    image: RoyalAlloyGP200LCRed,
+    colors: ["orange", "green"],
+    rate: 19,
+    Manufacturer: "Yamaha",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "125cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 9,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "48 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GT 150 AC",
+    price: "6529,99 zł",
+    image: UltraBlueIvory125cc,
+    colors: ["blue", "pink"],
+    rate: 23,
+    Manufacturer: "Forte",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "150cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 10,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "28 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GP 125 HX",
+    price: "6199,99 zł",
+    image: White2,
+    colors: ["purple", "blue"],
+    rate: 17,
+    Manufacturer: "Honda",
+    Country: "Japan",
+    "Power Type": "Petrol",
+    "Engine Capacity": "50cm3",
+    "Wheel size": "10 inches",
+    "Seat available": "1 person",
+    id: 11,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "34 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+  {
+    name: "GT 50 WP",
+    price: "3699,99 zł",
+    image: White,
+    colors: ["green", "blue"],
+    rate: 13,
+    Manufacturer: "Moto Tech",
+    Country: "China",
+    "Power Type": "Electric",
+    "Engine Capacity": "80cm3",
+    "Wheel size": "12 inches",
+    "Seat available": "2 person",
+    id: 12,
+    description:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Adipiscing sed urna neque ultrices. Hendrerit quam sit lacus ac at rhoncus pellentesque ac. Turpis facilisi justo tincidunt erat pellentesque sagittis amet etiam. Laoreet magna nec sit habitasse platea sit nunc. ",
+    "Top Speed": "37 mph",
+    "Delivery Date": [
+      {
+        country: "England",
+        time: "4-5 days",
+      },
+      {
+        country: "Australia",
+        time: "8-9 days",
+      },
+      {
+        country: "USA",
+        time: "11-14 days",
+      },
+      {
+        country: "Europe",
+        time: "3-4 days",
+      },
+    ],
+  },
+];
